add support for unix sockets, and any other arbitrary protocol supported by the net package
